Rename MemoryWritableStream.ts => streams.ts
