+++ app/server/lib/AttachmentStore.ts
@@ -2,7 +2,7 @@ import {
   ExternalStorage,
   joinKeySegments,
 } from 'app/server/lib/ExternalStorage';
-import {MemoryWritableStream} from 'app/server/utils/MemoryWritableStream';
+import {MemoryWritableStream} from 'app/server/utils/streams';
 import * as fse from 'fs-extra';
 import * as stream from 'node:stream';
 import * as path from 'path';
