+++ test/server/lib/ActiveDoc.ts
@@ -16,7 +16,7 @@ import log from 'app/server/lib/log';
 import {timeoutReached} from 'app/server/lib/serverUtils';
 import {Throttle} from 'app/server/lib/Throttle';
 import {createTmpDir as createTmpUploadDir, globalUploadSet} from 'app/server/lib/uploads';
-import {MemoryWritableStream} from 'app/server/utils/MemoryWritableStream';
+import {MemoryWritableStream} from 'app/server/utils/streams';
 import {promisify} from 'bluebird';
 import {assert} from 'chai';
 import decompress from 'decompress';
